+++ export/orbax/export/__init__.py
@@ -14,6 +14,8 @@
 
 """Defines exported symbols for Orbax Export."""
 
+# pylint: disable=g-importing-member
+from orbax.export import config
 from orbax.export import typing
 from orbax.export import utils
 from orbax.export.dtensor_utils import dtensor_initialized