+++ export/CHANGELOG.md
@@ -8,7 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
-- Add `orbax.export.typing` API
+- Add `orbax_export.config_class` API to set global configs.
+- Add `orbax_export.typing` API
 - Require jax version >=0.4.30 as the jax.export API was introduced in this release.